+++ Source/Scene/Scene.js
@@ -117,7 +117,7 @@ define([
             creditContainer.style.bottom = '0';
             creditContainer.style['text-shadow'] = '0px 0px 2px #000000';
             creditContainer.style.color = '#ffffff';
-            creditContainer.style['font-size'] = '10pt';
+            creditContainer.style['font-size'] = '10px';
             creditContainer.style['padding-right'] = '5px';
             canvas.parentNode.appendChild(creditContainer);
         }